+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
-## 5.1.2(未发版)
+## 5.1.2
 
 - @uiw-admin/config
   - 🌟 新增一个默认 `define` 的 `STORAGE` 属性,本地存储使用localStorage或sessionStorage ,默认 "session",可选值  local | session
@@ -16,6 +16,7 @@
   - 🌟 feat:新增`saveField`参数
 - @uiw-admin/components
   - 🌟 `PromForm`：增加SelectMultiple下拉多选组件
+  - 🌟 `PromTable`：增加列表选择框功能
   
 ## 5.1.2-alpha.1